敵弾速度調整、ファイル分けた
