+++ Objects/Boss/Bullet/BossNormalBullet.cpp
@@ -1,4 +1,4 @@
-#include "BossPillow.h"
+#include "BossNormalBullet.h"
 
 #include <ModelManager.h>
 #include <cmath>
@@ -6,7 +6,7 @@
 
 #include "CalculateMath.h"
 
-void BossPillow::Initialize()
+void BossNormalBullet::Initialize()
 {
 	// --- 3Dオブジェクト ---
 	ModelManager::GetInstance()->LoadModel("cube.obj");
@@ -15,24 +15,28 @@ void BossPillow::Initialize()
 	object_->Initialize("cube.obj");
 
 	// 仮置き
-	object_->SetSize({ 0.2f,0.2f,0.2f });
+	object_->SetSize({ 0.2f,0.2f,0.5f });
 
 
 	object_->SetPosition(position_);
 }
 
-void BossPillow::Finalize()
+void BossNormalBullet::Finalize()
 {
+	// 各解放処理
+	//isDead_ = true;
+	//object_.reset();
+
 	ModelManager::GetInstance()->Finalize();
 }
 
-void BossPillow::Update()
+void BossNormalBullet::Update()
 {
+
 	object_->Update();
 
 	object_->SetPosition(position_);
 	object_->SetRotate(rotation_);
-
 
 	// 敵弾から自キャラへのベクトルを計算
 	Vector3 toPlayer = playerPosition_ - position_;
@@ -44,8 +48,17 @@ void BossPillow::Update()
 	//velocity_ = 0.1f * Slerp(velocity_, toPlayer, 0.05f);
 	velocity_.Lerp(velocity_, toPlayer, 0.05f);
 
-	position_ += velocity_;
-	rotation_.y += 1.0f;
+	// 進行方向に見た目の回転を合わせる(ex1)
+	// Y軸周り角度(θy)
+	rotation_.y = std::atan2(velocity_.x, velocity_.z);
+	double velocityXZ = sqrt(pow(velocity_.x, 2) + pow(velocity_.z, 2));
+
+	// X軸周り角度(θx)
+	rotation_.x = (float)std::atan2(-velocity_.y, velocityXZ);
+
+	position_ += velocity_ * 0.1f;
+
+
 
 	//時間経過でデス
 	if (--deathTimer_ <= 0) {
@@ -53,7 +66,7 @@ void BossPillow::Update()
 	}
 }
 
-void BossPillow::Draw()
+void BossNormalBullet::Draw()
 {
 	object_->Draw();
 }
