+++ Objects/Boss/State/BossStateFourth.cpp
@@ -1,6 +1,6 @@
 #include "BossStateFourth.h"
 
-#include "./Objects/Boss/Boss.h"
+#include "../Boss.h"
 
 BossStateFourth::BossStateFourth(Boss* _pBoss) : BaseBossState("State Fourth", _pBoss)
 {
