+++ game.vcxproj.filters
@@ -60,37 +60,37 @@
     <ClCompile Include="Objects\Boss\Boss.cpp">
       <Filter>Object\Boss</Filter>
     </ClCompile>
-    <ClCompile Include="Player.cpp">
+    <ClCompile Include="Objects\Player\Player.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="PlayerBullet.cpp">
+    <ClCompile Include="Objects\Player\Bullet\PlayerBullet.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
     <ClCompile Include="Objects\Skydome\Skydome.cpp">
       <Filter>Object\Skydome</Filter>
     </ClCompile>
-    <ClCompile Include="BossNormalBullet.cpp">
+    <ClCompile Include="Objects\Boss\Bullet\BossNormalBullet.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="BaseBossState.cpp">
+    <ClCompile Include="Objects\Boss\State\BaseBossState.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="BossStateFirst.cpp">
+    <ClCompile Include="Objects\Boss\State\BossStateFirst.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="BossStateSecond.cpp">
+    <ClCompile Include="Objects\Boss\State\BossStateSecond.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="BossStateThird.cpp">
+    <ClCompile Include="Objects\Boss\State\BossStateThird.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="BossStateFourth.cpp">
+    <ClCompile Include="Objects\Boss\State\BossStateFourth.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="BossPillow.cpp">
+    <ClCompile Include="Objects\Boss\Bullet\BossPillow.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="BossMoon.cpp">
+    <ClCompile Include="Objects\Boss\Bullet\BossMoon.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
   </ItemGroup>
@@ -122,37 +122,37 @@
     <ClInclude Include="Objects\.GameObject\GameObject.h">
       <Filter>Object\GameObject</Filter>
     </ClInclude>
-    <ClInclude Include="Player.h">
+    <ClInclude Include="Objects\Player\Player.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="PlayerBullet.h">
+    <ClInclude Include="Objects\Player\Bullet\PlayerBullet.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
     <ClInclude Include="Objects\Skydome\Skydome.h">
       <Filter>Object\Skydome</Filter>
     </ClInclude>
-    <ClInclude Include="BossNormalBullet.h">
+    <ClInclude Include="Objects\Boss\Bullet\BossNormalBullet.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="BaseBossState.h">
+    <ClInclude Include="Objects\Boss\State\BaseBossState.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="BossStateFirst.h">
+    <ClInclude Include="Objects\Boss\State\BossStateFirst.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="BossStateSecond.h">
+    <ClInclude Include="Objects\Boss\State\BossStateSecond.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="BossStateThird.h">
+    <ClInclude Include="Objects\Boss\State\BossStateThird.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="BossStateFourth.h">
+    <ClInclude Include="Objects\Boss\State\BossStateFourth.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="BossPillow.h">
+    <ClInclude Include="Objects\Boss\Bullet\BossPillow.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
-    <ClInclude Include="BossMoon.h">
+    <ClInclude Include="Objects\Boss\Bullet\BossMoon.h">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
   </ItemGroup>
