+++ imgui.ini
@@ -4,7 +4,7 @@ Size=400,400
 Collapsed=0
 
 [Window][Easing Parameters]
-Pos=654,407
+Pos=831,104
 Size=500,600
 Collapsed=0
